+++ src/Drummer.h
@@ -46,7 +46,7 @@ class Drummer : public Instrmnt
     instrument.  An StkError will be thrown if the rawwave path is
     incorrectly set.
   */
-  void noteOn( StkFloat instrument, StkFloat amplitude );
+  void noteOn( StkFloat instrument, StkFloat amplitude, StkFloat frequency );
 
   //! Stop a note with the given amplitude (speed of decay).
   void noteOff( StkFloat amplitude );